requirements.txt
python-dotenv
google-genai
ffmpeg-python